Build script now uses streams.
The entire build process is a single stream, from start to finish.
It can scale to thousands or millions of lines of code, and never
run out of memory.
It can be redirected to the terminal or files by changing a single
line.
It can be piped into other commands, such as grep, on the command
line.